Check that tivoFile != null before attempting to delete
git-svn-id: https://svn.code.sf.net/p/kmttg/code/trunk@979 b0df3fdd-2057-4f73-b226-dfc9190e8f7a

+++ src/com/tivo/kmttg/task/qsfix.java
@@ -207,7 +207,7 @@ public class qsfix {
             log.print("---DONE--- job=" + job.type + " output=" + job.mpegFile_fix);
             
             // Remove .TiVo file if option enabled
-            if (config.RemoveTivoFile == 1 && config.VrdDecrypt == 1) {
+            if (job.tivoFile != null && config.RemoveTivoFile == 1 && config.VrdDecrypt == 1) {
                if ( file.delete(job.tivoFile) ) {
                   log.print("(Deleted file: " + job.tivoFile + ")");
                } else {
